added conditional for if item is not in te other join table

+++ app/controllers/application_controller.rb
@@ -16,10 +16,18 @@ class ApplicationController < ActionController::Base
   def clean_join_tables
     if /inventories/.match(params[:controller])
       @item = Item.find_by(name: params[:inventory][:items_attributes][:'0'][:name])
-      WishlistItem.where(item_id: @item.id).destroy_all
+      if @item != nil
+        if WishlistItem.where(item_id: @item.id)
+          WishlistItem.where(item_id: @item.id).destroy_all
+        end
+      end
     elsif /wishlists/.match(params[:controller])
-      @item = Item.find_by(name: params[:inventory][:items_attributes][:'0'][:name])
-      InventoryItem.where(item_id: @item.id).destroy_all
+      @item = Item.find_by(name: params[:wishlist][:items_attributes][:'0'][:name])
+      if @item != nil
+        if InventoryItem.where(item_id: @item.id)
+          InventoryItem.where(item_id: @item.id).destroy_all
+        end
+      end
     end
   end
 end
